Create new audit role to setup rsyslog and fix user account issues.
This commit creates a new role within this ansible repo, allowing a
provisioned machine to forward a set of logs onto a centralised rsyslogd
server. This can be illustrated and tested by the additional changes to
the vagrantfile to test the correct routing of messages to the server.

We also fix a number of related security issues:

 * Increase password strength (we don't ever use passwords, only ssh key pairs)
 * Increase the failed login block to restrict brute force attacks.
 * Add PAM setup to log and forward all user access and escalations to rsyslog.